integrated spin-cast animation into the photoresist button, added ability for waifer to spin
