+++ Assets/3D Assets/Play Area Spin.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c4bd3dc086472d448a7deb038b601c7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
