+++ Assets/Scripts/animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b56fe127a260587418bbec74dec5c89a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
