--- Assets/Scripts/animationDestroy.cs
@@ -1,25 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class animationDestroy : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void destroyMe()
-    {
-        GameObject parent = gameObject.transform.parent.gameObject;
-        Destroy(parent);
-        Destroy(gameObject);
-    }
-}
