+++ Current Controls.txt
@@ -2,4 +2,3 @@ Controls:
 
 -Space Bar: Photomask liftoff
 -L: Check level requirements
--P: Create a spin caster animation
