+++ Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,4 +1,3 @@
 Base path: 'C:/Program Files/Unity/Hub/Editor/2021.3.13f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2021.3.13f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
-Cmd: shutdown
